backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Base de datos
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# PostGIS / Geolocalización
geoalchemy2==0.15.2
shapely==2.0.6

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validación
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Utilidades
python-dotenv==1.0.1
httpx==0.27.2

# Servicios externos
stripe==10.12.0
firebase-admin==6.5.0
cloudinary==1.41.0
twilio==9.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
